requirements.txt
pdfservices-sdk==2.2.0
Pillow==9.5.0
pdfrw==0.4
opencv-python==4.7.0.72
pdf2image==1.16.3
pandas==2.0.2
langchain==0.0.198
python-dotenv==1.0.0
openai==0.27.8
